move forum into view
