+++ packages/client/src/components/Forum/ThemePage.tsx
@@ -2,8 +2,8 @@ import React, { useEffect, useRef } from 'react'
 import { useNavigate, useParams, useLocation } from 'react-router-dom'
 import { useForm, SubmitHandler, FieldValues } from 'react-hook-form'
 import cn from './ThemePage.module.scss'
-import { Message } from './types'
-import { messages, topics } from './forumData'
+import { Message } from '../../views/Forum/types'
+import { messages, topics } from '../../views/Forum/forumData'
 import Button from '../UI/Button'
 import Input from '../UI/Input'
 import removePathSuffix from '../../utils/removePathSuffix'
